#38
- реструктуризация раздела preset (major)
- PresetBrowser переименован в PresetWeb (major)
- SpecificExtras: meta переименован в setMeta (major)
