+++ src/preset/node/node/utils/http-client/__test__/index.test.ts
@@ -1,70 +1,9 @@
 import { Context } from '@opentelemetry/api';
 import { SemanticAttributes } from '@opentelemetry/semantic-conventions';
 import { Span, Tracer } from '@opentelemetry/sdk-trace-base';
-import { tracingMiddleware, getRequestInfo, hideFirstId, displayUrl } from '..';
-
-describe('displayUrl', () => {
-  it('should properly merge baseURL with url', () => {
-    const cases: Array<{ url: string; baseURL: string; expectedUrl: string }> = [
-      // baseURL и url
-      {
-        baseURL: 'https://www.base.com/',
-        url: '/user/current',
-        expectedUrl: 'https://www.base.com/user/current',
-      },
-      {
-        baseURL: 'https://www.base.com',
-        url: 'user/current',
-        expectedUrl: 'https://www.base.com/user/current',
-      },
-      {
-        baseURL: 'https://www.base.com',
-        url: '/user/current',
-        expectedUrl: 'https://www.base.com/user/current',
-      },
-      {
-        baseURL: 'https://www.base.com/',
-        url: 'admin/all',
-        expectedUrl: 'https://www.base.com/admin/all',
-      },
-
-      // только baseURL
-      {
-        baseURL: 'www.test.com',
-        url: '',
-        expectedUrl: 'www.test.com',
-      },
-      {
-        baseURL: 'www.test.com/',
-        url: '',
-        expectedUrl: 'www.test.com/',
-      },
-
-      // только url
-      {
-        baseURL: '',
-        url: '/hello/world',
-        expectedUrl: '/hello/world',
-      },
-      {
-        baseURL: '',
-        url: 'some/path',
-        expectedUrl: 'some/path',
-      },
-
-      // ничего
-      {
-        baseURL: '',
-        url: '',
-        expectedUrl: '[empty]',
-      },
-    ];
-
-    for (const { baseURL, url, expectedUrl } of cases) {
-      expect(displayUrl(baseURL, url)).toBe(expectedUrl);
-    }
-  });
-});
+import { tracingMiddleware, getRequestInfo, hideFirstId, getRequestHeaders } from '..';
+import { BaseConfig } from '../../../../../../config';
+import { Request } from 'express';
 
 describe('tracingMiddleware', () => {
   it('should handle success response', async () => {
@@ -224,6 +163,70 @@ describe('hideFirstId', () => {
   it('should replace first id only', () => {
     cases.forEach(({ input, output }) => {
       expect(hideFirstId(input)).toEqual(output);
+    });
+  });
+});
+
+describe('getRequestHeaders', () => {
+  it('should return headers', () => {
+    const config: BaseConfig = {
+      appName: 'foo',
+      appVersion: '0.0.1',
+      env: 'test',
+    };
+
+    const request: Request = {
+      socket: {
+        remoteAddress: '127.0.0.1',
+      },
+      headers: {
+        cookie: 'userid=12345',
+        'simaland-a': 'aaa',
+        'simaland-b': 'bbb',
+      },
+      get(key: string) {
+        return this.headers[key];
+      },
+      header(key: string) {
+        return this.headers[key];
+      },
+    } as any;
+
+    const result = getRequestHeaders(config, request);
+
+    expect(result).toEqual({
+      'X-Client-Ip': '127.0.0.1',
+      'User-Agent': `simaland-foo/0.0.1`,
+      Cookie: 'userid=12345',
+      'simaland-a': 'aaa',
+      'simaland-b': 'bbb',
+    });
+  });
+
+  it('should return headers when lot of data is undefined', () => {
+    const config: BaseConfig = {
+      appName: 'foo',
+      appVersion: '0.0.1',
+      env: 'test',
+    };
+
+    const request: Request = {
+      socket: {
+        remoteAddress: undefined,
+      },
+      headers: {},
+      get(key: string) {
+        return this.headers[key];
+      },
+      header(key: string) {
+        return this.headers[key];
+      },
+    } as any;
+
+    const result = getRequestHeaders(config, request);
+
+    expect(result).toEqual({
+      'User-Agent': `simaland-foo/0.0.1`,
     });
   });
 });
